fix bug rpcapi and add debug
