+++ lib/api/index.js
@@ -1,5 +1,5 @@
 'use strict'
-const debugDdvApi = require('debug')('api/index')
+const debug = require('debug')('api/index')
 // ws模块
 const Ws = require('ws')
 // 工具模块
@@ -16,15 +16,15 @@ const config = require(ddvPath('lib/config'))
 const Api = module.exports = class DdvApi extends domain.Domain {
   // 构造函数
   constructor (o) {
-    debugDdvApi('new')
+    debug('new')
     // 调用父类构造函数
     super()
     // 时间
     this.startTimeStamp = util.now() / 1000
     // 基本数据初始化
     this._baseInit(o)
-    debugDdvApi('new address:', this.address)
-    debugDdvApi('new rpcTimeout:', this.rpcTimeout)
+    debug('new address:', this.address)
+    debug('new rpcTimeout:', this.rpcTimeout)
     // 初始化响应事件
     this._requestResponseEvnetInit()
     // 返回实例化对象或者开始连接
@@ -52,9 +52,9 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   // 连接
   connect () {
     // 调试
-    debugDdvApi('connect address:', this.address)
+    debug('connect address:', this.address)
     // 调试
-    debugDdvApi('connect rpcTimeout:', this.rpcTimeout)
+    debug('connect rpcTimeout:', this.rpcTimeout)
     // 返回一个Promise
     return new Promise((resolve, reject) => {
       // 建立客户端
@@ -73,6 +73,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
       this._reject = reject
       // 设定超时
       this.openWsSetTimeout = setTimeout(() => {
+        debug('rpc api connect timeout:', this.address)
         // 回调错误 如果还没有连接成功就超时
         this._reject && this._reject(new Error('rpc api connect timeout'))
       }, this.rpcTimeout)
@@ -96,7 +97,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   // rpc-api-ws 连接成功事件
   _eventOpen () {
     // 调试
-    debugDdvApi('ws evnet open address:' + this.address)
+    debug('ws evnet open address:' + this.address)
     // 触发wsopen
     this.emit('wsopen')
     // ws连接成功
@@ -108,6 +109,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
       this.state = true
       this.pid = pid
       this._resolve && this._resolve(this)
+      debug('rpcapi evnet open address:' + this.address)
       this.emit('open')
     }).catch(e => {
       // 防止定时器占用事件循环队列
@@ -119,7 +121,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   }
   _eventClose () {
     // 调试
-    debugDdvApi('ws event close address:' + this.address)
+    debug('ws event close address:' + this.address)
     // 重置状态
     this._resetWsState()
     // 触发关闭事件
@@ -128,7 +130,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
     // rpc-api-ws 出错事件
   _eventError (e) {
     // 调试
-    debugDdvApi('ws event error address:' + this.address)
+    debug('ws event error address:' + this.address)
     // 重置状态
     this._resetWsState()
     // 触发关闭事件
@@ -138,6 +140,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   }
     // rpc-api-ws 消息事件
   _eventMessage (body) {
+    debug('ws event _eventMessage', body)
     var res
     try {
       res = rowraw.parse(body)
@@ -148,17 +151,17 @@ const Api = module.exports = class DdvApi extends domain.Domain {
       if (res.type === 'request') {
         if (res.protocol) {
           if (!(res.protocol && this.emit(('protocol::' + res.protocol.toLowerCase()), res))) {
-            debugDdvApi('protocol::notfind', res)
+            debug('protocol::notfind', res)
           }
         } else {
-          debugDdvApi('protocol::other', res)
+          debug('protocol::other', res)
         }
       } else if (res.type === 'response') {
         if (!(res.protocol && this.emit('request::response', res))) {
-          debugDdvApi('response::notfind', res)
+          debug('response::notfind', res)
         }
       } else {
-        debugDdvApi('response::other', res)
+        debug('response::other', res)
       }
     }
   }
@@ -184,11 +187,13 @@ const Api = module.exports = class DdvApi extends domain.Domain {
     this.on('request::response', this._requestResponse.bind(this))
   }
   _requestResponse (rowraw) {
+    debug('request::response-_requestResponse', rowraw)
     var requestId, o, code, e
     if (!(rowraw.headers && (requestId = rowraw.headers.request_id))) {
       return
     }
     if (this.processWs && (o = this.processWs[requestId])) {
+      debug('request::response-_requestResponse-processWs', requestId, o)
       code = parseInt(rowraw.status || 0) || 0
       if (code >= 200 && code < 300) {
         if (util.isFunction(o.resolve)) {
@@ -209,12 +214,11 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   request (o) {
     return new Promise((resolve, reject) => {
       o.headers = o.headers || Object.create(null)
-      o.requestId = o.headers.request_id || o.requestId || util.createRequestId()
-      o.headers.requestId = o.requestId
+      o.headers.request_id = o.requestId = o.requestId || o.headers.request_id || util.createRequestId()
       o.body = o.body || ''
       o.start = o.start || ''
-      o.resolve = o.resolve
-      o.reject = o.reject
+      o.resolve = resolve
+      o.reject = reject
       o.addtime = util.time()
       o.raw = rowraw.stringify(o.headers, o.body, o.start)
       delete o.headers
@@ -239,6 +243,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
   }
   // 发送命令行数据
   cmd (command, argv) {
+    debug('cmd', command, argv)
     return this.request({
       'start': 'COMMAND /v1_0/command/' + command + ' RPCAPI/1.0',
       'headers': {
@@ -247,6 +252,7 @@ const Api = module.exports = class DdvApi extends domain.Domain {
     })
   }
   ping () {
+    debug('ws ping address:' + this.address)
     return this.request({
       'start': 'PING /v1_0/init RPCAPI/1.0'
     }).then(({headers}) => {
