+++ lib/master/which.js
@@ -1,6 +1,8 @@
 // NpmWhich 查看命令所在目录
 const NpmWhich = require('npm-which')
+// 路径模块
 const path = require('path')
+// 文件系统模块
 const fs = require('fs')
 // 导出模块
 const which = module.exports = function which (commandDir) {
