+++ lib/server/command.js
@@ -1,4 +1,5 @@
 'use strict'
+const debug = require('debug')('server/command')
 // ddv路径模块
 const ddvPath = require('../ddvPath')
 // 配置项
@@ -11,6 +12,7 @@ const siteManage = require(ddvPath('lib/site/manage.js'))
 const Rowraw = require('ddv-rowraw')
 // 通过头中提取argv
 const getCommandArgvByHeaders = (headers) => {
+  debug('getCommandArgvByHeaders', headers)
   var argv = null
   if (headers && typeof headers === 'object' && headers.command_argv) {
     try {
@@ -48,6 +50,7 @@ const getRowrawByErrOrBody = (headers, err, body) => {
 }
 // 通过头中提取argv
 const checkHeadersRequestId = (headers) => {
+  debug('checkHeadersRequestId', headers)
   if (headers && typeof headers === 'object' && headers.request_id) {
     headers = void 0
     return true
@@ -57,8 +60,10 @@ const checkHeadersRequestId = (headers) => {
 }
 // 绑定事件
 const commandEventBind = function commandEventBind (WebSocket) {
+  debug('commandEventBind')
   // rpcapi事件转发
   WebSocket.on('protocol::rpcapi', (rowraw) => {
+    debug('protocol::rpcapi', rowraw)
     if (!(rowraw.method && rowraw.path && this.emit(['rpcapi', rowraw.method.toLowerCase(), rowraw.path], rowraw.headers, rowraw.body, rowraw))) {
       let stat = 'RPCAPI/1.0 404 Not Found'
       let raw = Rowraw.stringify({
@@ -70,6 +75,7 @@ const commandEventBind = function commandEventBind (WebSocket) {
   })
   // ping
   WebSocket.on(['rpcapi', 'ping', '/v1_0/init'], (headers) => {
+    debug('protocol::rpcapi::ping::/v1_0/init', headers)
     let stat = 'RPCAPI/1.0 200 OK'
     let raw = Rowraw.stringify({
       'request_id': headers.request_id,
