+++ lib/server/rpcapi.js
@@ -1,4 +1,5 @@
 'use strict'
+const debug = require('debug')('server/rpcapi')
 const ddvPath = require('../ddvPath')
 // 对外api
 const api = module.exports = Object.create(null)
@@ -106,6 +107,7 @@ fn.connectionWs = function serverConnectionWs (WebSocket) {
     }
   })
   WebSocket.on('message', function _onWsMessage (body) {
+    debug('WebSocket - message')
     if (!this.notClosed) { return console.error(Error('socket close'))/* 防止客户断开连接后继续有程序运行 */ }
     var type, res
     try {
@@ -114,8 +116,10 @@ fn.connectionWs = function serverConnectionWs (WebSocket) {
       res = null
     }
     if (res) {
+      debug('WebSocket - message - res', res)
       if (res.type === 'request') {
         if (res.protocol) {
+          debug('WebSocket - message - res - protocol', res.protocol)
           if (!(res.protocol && this.emit(('protocol::' + res.protocol.toLowerCase()), res))) {
             this.send(('not find protocol:' + (res.protocol || '')), function sendCb () {})
           }
